set up routes. added functions.js module. fixed FK dropdown logic. fixed Characters table JOIN query.
