+++ routes/index.js
@@ -1,0 +1,8 @@
+var express = require('express');
+var router = express.Router();
+
+router.get('/', function(req, res, next){
+    res.render('index.ejs');
+});
+
+module.exports = router;
