+++ views/addRemoveCharacters.ejs
@@ -1,5 +1,5 @@
 <!-- New Character Form -->
-<form action="/addRemoveCharacters?action=add" method="post">
+<form action="/characters/addRemoveCharacters?action=add" method="post">
     <fieldset class="form-control">
         <legend>Add Character</legend>
         <div class="row mb-3">
@@ -52,7 +52,7 @@
 
 <!-- Remove Character Form -->
 
-<form action="/addRemoveCharacters?action=delete" method="post">
+<form action="/characters/addRemoveCharacters?action=delete" method="post">
     <fieldset class="form-control">
         <legend>Remove Character</legend>
         <div class="row mb-3">
