+++ views/addRemoveSpells.ejs
@@ -1,6 +1,6 @@
 <!-- New Spell Form -->
 
-<form action="addRemoveSpells?action=add" method="post">
+<form action="/spells/addRemoveSpells?action=add" method="post">
     <fieldset class="form-control">
         <legend>Add Spell</legend>
         <div class="row mb-3">
@@ -41,7 +41,7 @@
 
 <!-- Remove Spell Form -->
 
-<form action="/addRemoveSpells?action=delete" method="post">
+<form action="/spells/addRemoveSpells?action=delete" method="post">
     <fieldset class="form-control">
         <legend>Remove Spell</legend>
         <div class="row mb-3">
