+++ views/alterCharacters.ejs
@@ -1,6 +1,6 @@
 <!-- Alter Character Form -->
 
-<form action="/alterCharacters" method="post">
+<form action="/characters/alterCharacters" method="post">
     <fieldset class="form-control">
         <legend>Alter Character</legend>
         <div class="row mb-3">
@@ -66,7 +66,7 @@
     document.getElementById('charID').addEventListener('change', async function (event) {
         var payload = {}
         payload.charID = document.getElementById('charID').value;
-        fetch('/alterCharacters/autofill', {
+        fetch('/characters/alterCharacters/autofill', {
             method: 'POST',
             headers: {
                 'Content-Type': 'application/json'
@@ -91,7 +91,7 @@
     function autofill() {
         var payload = {}
         payload.charID = document.getElementById('charID').value;
-        fetch('/alterCharacters/autofill', {
+        fetch('/characters/alterCharacters/autofill', {
             method: 'POST',
             headers: {
                 'Content-Type': 'application/json'
