+++ views/alterItems.ejs
@@ -7,7 +7,7 @@
                 <label for="itemID" class="form-label">Item ID:</label>
                 <select class="form-select" id="itemID">
                   <% itemIDs.forEach(function(item) { %>
-                    <option> <%= item.itemID %>: <%=item.name %> </option>
+                    <option value="<%= item.itemID %>"> <%= item.itemID %>: <%=item.name %> </option>
                   <% }) %>
                 </select>
             </div>
@@ -40,11 +40,11 @@ document.getElementById('button').addEventListener('click', async function(event
             var req = new XMLHttpRequest();
             // CREATING JSON OBJECT TO SEND AS POST REQUEST
             var payload = {id:null, name:null, damage:null, cost:null};
-            payload.id = document.getElementById('itemID').value[0];
+            payload.id = document.getElementById('itemID').value;
             payload.name = document.getElementById('name').value;
             payload.damage = document.getElementById('damage').value;
             payload.cost = document.getElementById('cost').value
-            req.open('POST', '/alterItems');
+            req.open('POST', '/items/alterItems');
             req.setRequestHeader('Content-Type', 'application/json');
 
             // PARSING RESPONSE SENT FROM SERVER
@@ -61,10 +61,10 @@ document.getElementById('button').addEventListener('click', async function(event
 // MAKE GET REQUEST TO POPULATE EDIT FORM VALUES
 document.getElementById('itemID').addEventListener('change', async function(event){
     console.log('changed menu')
-            let id = document.getElementById('itemID').value[0]
+            let id = document.getElementById('itemID').value
             console.log(id)
             var req = new XMLHttpRequest();
-            req.open('GET', '/alterItems?id=' + id);
+            req.open('GET', '/items/alterItems?id=' + id);
             req.addEventListener('load', function(event) {
                 if (req.status >= 200 && req.status < 400) {
                     var response = JSON.parse(req.responseText);
