+++ views/index.ejs
@@ -9,11 +9,11 @@
         <h5 class="card-title">Characters</h5>
         <ul class="list-unstyled">
             <li>Browse and search existing Characters</li>
-            <a href="characters"><button type="button" class="btn btn-primary btn-sm mb-3" >View Characters</button></a>
+            <a href="/characters"><button type="button" class="btn btn-primary btn-sm mb-3" >View Characters</button></a>
             <li>Add new Characters, or remove existing Characters</li>
-            <a href="addRemoveCharacters"><button type="button" class="btn btn-primary btn-sm mb-3" >Add/Remove Characters</button></a>
+            <a href="/characters/addRemoveCharacters"><button type="button" class="btn btn-primary btn-sm mb-3" >Add/Remove Characters</button></a>
             <li>Update existing Characters</li>
-            <a href="alterCharacters"><button type="button" class="btn btn-primary btn-sm mb-3" >Alter Characters</button></a>
+            <a href="/characters/alterCharacters"><button type="button" class="btn btn-primary btn-sm mb-3" >Alter Characters</button></a>
         </ul>
       </div>
     </div>
@@ -25,11 +25,11 @@
         <h5 class="card-title">Spells</h5>
         <ul class="list-unstyled">
             <li>Browse and search existing Spells</li>
-            <a href="spells"><button type="button" class="btn btn-secondary btn-sm mb-3" >View Spells</button></a>
+            <a href="/spells"><button type="button" class="btn btn-secondary btn-sm mb-3" >View Spells</button></a>
             <li>Add new Spells, or remove existing Spells</li>
-            <a href="addRemoveSpells"><button type="button" class="btn btn-secondary btn-sm mb-3" >Add/Remove Spells</button></a>
+            <a href="/spells/addRemoveSpells"><button type="button" class="btn btn-secondary btn-sm mb-3" >Add/Remove Spells</button></a>
             <li>Update existing Spells</li>
-            <a href="alterSpells"><button type="button" class="btn btn-secondary btn-sm mb-3" >Alter Spells</button></a>
+            <a href="/spells/alterSpells"><button type="button" class="btn btn-secondary btn-sm mb-3" >Alter Spells</button></a>
         </ul>
       </div>
     </div>
@@ -41,13 +41,13 @@
         <h5 class="card-title">Items</h5>
         <ul class="list-unstyled">
             <li>Browse and search existing Items</li>
-            <a href="items"><button type="button" class="btn btn-success btn-sm mb-3" >Items</button></a>
+            <a href="/items"><button type="button" class="btn btn-success btn-sm mb-3" >Items</button></a>
             <li>Add new Items, or remove existing Items</li>
-            <a href="addRemoveItems"><button type="button" class="btn btn-success btn-sm mb-3" >Add/Remove Items</button></a>
+            <a href="/items/addRemoveItems"><button type="button" class="btn btn-success btn-sm mb-3" >Add/Remove Items</button></a>
             <li>Update existing Items</li>
-            <a href="alterItems"><button type="button" class="btn btn-success btn-sm mb-3" >Alter Items</button></a>
+            <a href="/items/alterItems"><button type="button" class="btn btn-success btn-sm mb-3" >Alter Items</button></a>
             <li>Assign Items to Characters</li>
-            <a href="characterItems"><button type="button" class="btn btn-success btn-sm mb-3" >Items to Characters</button></a>
+            <a href="/items/characterItems"><button type="button" class="btn btn-success btn-sm mb-3" >Items to Characters</button></a>
         </ul>
       </div>
     </div>
@@ -59,11 +59,11 @@
         <h5 class="card-title">Regions</h5>
         <ul class="list-unstyled">
             <li>Browse and search existing Regions</li>
-            <a href="regions"><button type="button" class="btn btn-info btn-sm mb-3" >Regions</button></a>
+            <a href="/regions"><button type="button" class="btn btn-info btn-sm mb-3" >Regions</button></a>
             <li>Add new Regions, or remove existing Regions</li>
-            <a href="addRemoveRegions"><button type="button" class="btn btn-info btn-sm mb-3" >Add/Remove Regions</button></a>
+            <a href="/regions/addRemoveRegions"><button type="button" class="btn btn-info btn-sm mb-3" >Add/Remove Regions</button></a>
             <li>Update existing Regions</li>
-            <a href="alterRegions"><button type="button" class="btn btn-info btn-sm mb-3" >Alter Regions</button></a>
+            <a href="/regions/alterRegions"><button type="button" class="btn btn-info btn-sm mb-3" >Alter Regions</button></a>
         </ul>
       </div>
     </div>
@@ -75,13 +75,13 @@
         <h5 class="card-title">Enemies</h5>
         <ul class="list-unstyled">
             <li>Browse and search existing Enemies</li>
-            <a href="enemies"><button type="button" class="btn btn-dark btn-sm mb-3" >Enemies</button></a>
+            <a href="/enemies"><button type="button" class="btn btn-dark btn-sm mb-3" >Enemies</button></a>
             <li>Add new Enemies, or remove existing Enemies</li>
-            <a href="addRemoveEnemies"><button type="button" class="btn btn-dark btn-sm mb-3">Add/Remove Enemies</button></a>
+            <a href="/enemies/addRemoveEnemies"><button type="button" class="btn btn-dark btn-sm mb-3">Add/Remove Enemies</button></a>
             <li>Update existing Enemies</li>
-            <a href="alterEnemies"><button type="button" class="btn btn-dark btn-sm mb-3">Alter Enemies</button></a>
+            <a href="/enemies/alterEnemies"><button type="button" class="btn btn-dark btn-sm mb-3">Alter Enemies</button></a>
             <li>Assign Enemies to Regions</li>
-            <a href="regionEnemies"><button type="button" class="btn btn-dark btn-sm mb-3">Regions to Enemies</button></a>
+            <a href="/enemies/regionEnemies"><button type="button" class="btn btn-dark btn-sm mb-3">Regions to Enemies</button></a>
         </ul>
       </div>
     </div>
