+++ views/layout.ejs
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1">
-    <title>Nuclear Jalapenos Index</title>
+    <title>Nuclear Jalapenos</title>
 
     <!-- Bootstrap CSS -->
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-eOJMYsd53ii+scO/bJGFsiCZc+5NDVN2yr8+0RDqr0Ql0h+rP48ckxlpbzKgwra6" crossorigin="anonymous">
@@ -20,23 +20,23 @@
         <div class="collapse navbar-collapse" id="navbarNav">
             <ul class="navbar-nav">
                 <li class="nav-item dropdown">
-                    <a class="nav-link dropdown-toggle" href="characters" id="navbarDropdownMenuLink" role="button" data-bs-toggle="dropdown" aria-expanded="false">
+                    <a class="nav-link dropdown-toggle" href="/characters" id="navbarDropdownMenuLink" role="button" data-bs-toggle="dropdown" aria-expanded="false">
                         Characters
                     </a>
                     <ul class="dropdown-menu" aria-labelledby="navbarDropdownMenuLink">
-                        <li><a class="dropdown-item" aria-current="page" href="characters">Characters</a></li>
-                        <li><a class="dropdown-item" href="addRemoveCharacters">Add/Remove Characters</a></li>
-                        <li><a class="dropdown-item" href="alterCharacters">Alter Characters</a></li>
+                        <li><a class="dropdown-item" aria-current="page" href="/characters">Characters</a></li>
+                        <li><a class="dropdown-item" href="/characters/addRemoveCharacters">Add/Remove Characters</a></li>
+                        <li><a class="dropdown-item" href="/characters/alterCharacters">Alter Characters</a></li>
                     </ul>
                 </li>
                 <li class="nav-item dropdown">
-                    <a class="nav-link dropdown-toggle" href="spells" id="navbarDropdownMenuLink2" role="button" data-bs-toggle="dropdown" aria-expanded="false">
+                    <a class="nav-link dropdown-toggle" href="/spells" id="navbarDropdownMenuLink2" role="button" data-bs-toggle="dropdown" aria-expanded="false">
                         Spells
                     </a>
                     <ul class="dropdown-menu" aria-labelledby="navbarDropdownMenuLink2">
-                        <li><a class="dropdown-item" aria-current="page" href="spells">Spells</a></li>
-                        <li><a class="dropdown-item" href="addRemoveSpells">Add/Remove Spells</a></li>
-                        <li><a class="dropdown-item" href="alterSpells">Alter Spells</a></li>
+                        <li><a class="dropdown-item" aria-current="page" href="/spells">Spells</a></li>
+                        <li><a class="dropdown-item" href="/spells/addRemoveSpells">Add/Remove Spells</a></li>
+                        <li><a class="dropdown-item" href="/spells/alterSpells">Alter Spells</a></li>
                     </ul>
                 </li>
                 <li class="nav-item dropdown">
@@ -44,31 +44,31 @@
                         Items
                     </a>
                     <ul class="dropdown-menu" aria-labelledby="navbarDropdownMenuLink3">
-                        <li><a class="dropdown-item" aria-current="page" href="items">Items</a></li>
-                        <li><a class="dropdown-item" href="addRemoveItems">Add/Remove Items</a></li>
-                        <li><a class="dropdown-item" href="alterItems">Alter Items</a></li>
-                        <li><a class="dropdown-item" href="characterItems">Assign Items to Characters</a></li>
+                        <li><a class="dropdown-item" aria-current="page" href="/items">Items</a></li>
+                        <li><a class="dropdown-item" href="/items/addRemoveItems">Add/Remove Items</a></li>
+                        <li><a class="dropdown-item" href="/items/alterItems">Alter Items</a></li>
+                        <li><a class="dropdown-item" href="/items/characterItems">Assign Items to Characters</a></li>
                     </ul>
                 </li>
                 <li class="nav-item dropdown">
-                    <a class="nav-link dropdown-toggle" href="regions" id="navbarDropdownMenuLink4" role="button" data-bs-toggle="dropdown" aria-expanded="false">
+                    <a class="nav-link dropdown-toggle" href="/regions" id="navbarDropdownMenuLink4" role="button" data-bs-toggle="dropdown" aria-expanded="false">
                         Regions
                     </a>
                     <ul class="dropdown-menu" aria-labelledby="navbarDropdownMenuLink4">
-                        <li><a class="dropdown-item" aria-current="page" href="regions">Regions</a></li>
-                        <li><a class="dropdown-item" href="addRemoveRegions">Add/Remove Regions</a></li>
-                        <li><a class="dropdown-item" href="alterRegions">Alter Regions</a></li>
+                        <li><a class="dropdown-item" aria-current="page" href="/regions">Regions</a></li>
+                        <li><a class="dropdown-item" href="/regions/addRemoveRegions">Add/Remove Regions</a></li>
+                        <li><a class="dropdown-item" href="/regions/alterRegions">Alter Regions</a></li>
                     </ul>
                 </li>
                 <li class="nav-item dropdown">
-                    <a class="nav-link dropdown-toggle" href="enemies" id="navbarDropdownMenuLink5" role="button" data-bs-toggle="dropdown" aria-expanded="false">
+                    <a class="nav-link dropdown-toggle" href="/enemies" id="navbarDropdownMenuLink5" role="button" data-bs-toggle="dropdown" aria-expanded="false">
                         Enemies
                     </a>
                     <ul class="dropdown-menu" aria-labelledby="navbarDropdownMenuLink5">
-                        <li><a class="dropdown-item" aria-current="page" href="enemies">Enemies</a></li>
-                        <li><a class="dropdown-item" href="addRemoveEnemies">Add/Remove Enemies</a></li>
-                        <li><a class="dropdown-item" href="alterEnemies">Alter Enemies</a></li>
-                        <li><a class="dropdown-item" href="RegionEnemies">Assign Enemies to Regions</a></li>
+                        <li><a class="dropdown-item" aria-current="page" href="/enemies">Enemies</a></li>
+                        <li><a class="dropdown-item" href="/enemies/addRemoveEnemies">Add/Remove Enemies</a></li>
+                        <li><a class="dropdown-item" href="/enemies/alterEnemies">Alter Enemies</a></li>
+                        <li><a class="dropdown-item" href="/enemies/regionEnemies">Assign Enemies to Regions</a></li>
                     </ul>
                 </li>
             </ul>
