+++ views/regionEnemies.ejs
@@ -7,7 +7,7 @@
                 <label for="enemyID" class="form-label">Enemy ID:</label>
                 <select class="form-select" id="enemyID">
                   <% enemyIDs.forEach(function(enemy) { %>
-                    <option> <%= enemy.enemyID %>: <%= enemy.name %> </option>
+                    <option value="<%= enemy.enemyID %>"> <%= enemy.enemyID %>: <%= enemy.name %> </option>
                  <% }) %>
                 </select>
             </div>
@@ -15,7 +15,7 @@
                 <label for="regionID" class="form-label">Region ID:</label>
                 <select class="form-select" id="regionID">
                   <% regionIDs.forEach(function(region) { %>
-                    <option> <%= region.regionID %>: <%= region.name %> </option>
+                    <option value="<%= region.regionID %>"> <%= region.regionID %>: <%= region.name %> </option>
                  <% }) %>
                 </select>
             </div>
@@ -37,7 +37,7 @@
                 <label for="enemyID" class="form-label">Enemy ID:</label>
                 <select class="form-select" id="removeenemyID">
                   <% enemyIDs.forEach(function(enemy) { %>
-                    <option> <%= enemy.enemyID %>: <%= enemy.name %> </option>
+                    <option value="<%= enemy.enemyID %>"> <%= enemy.enemyID %>: <%= enemy.name %> </option>
                  <% }) %>
                 </select>
             </div>
@@ -45,7 +45,7 @@
                 <label for="regionID" class="form-label">Region ID:</label>
                 <select class="form-select" id="removeregionID">
                   <% regionIDs.forEach(function(region) { %>
-                    <option> <%= region.regionID %>: <%= region.name %> </option>
+                    <option value="<%= region.regionID %>"> <%= region.regionID %>: <%= region.name %> </option>
                  <% }) %>
                 </select>
             </div>
@@ -62,7 +62,7 @@
 <div class="container-fluid">
     <div class="row">
         <div class="col-3">
-            <form method="POST" action="/RegionEnemies?regionSearch=true">
+            <form method="POST" action="/regions/regionEnemies?regionSearch=true">
                 <fieldset>
                     <div class="row mb-3">
                         <div class="col-6">
@@ -78,7 +78,7 @@
 
 <!-- Enemy Search Bar -->
         <div class="col-6">
-            <form method="POST" action="/RegionEnemies?enemySearch=true">
+            <form method="POST" action="/regions/regionEnemies?enemySearch=true">
                 <fieldset>
                     <div class="row mb-3">
                         <div class="col-4">
@@ -183,9 +183,9 @@ document.getElementById('addbutton').addEventListener('click', async function(ev
     var req = new XMLHttpRequest();
     // CREATING JSON OBJECT TO SEND AS POST REQUEST
     var payload = {enemyID: null, regionID: null};
-    payload.enemyID = document.getElementById('enemyID').value[0];
-    payload.regionID = document.getElementById('regionID').value[0];
-    req.open('POST', '/RegionEnemies?action=add');
+    payload.enemyID = document.getElementById('enemyID').value;
+    payload.regionID = document.getElementById('regionID').value;
+    req.open('POST', '/regions/regionEnemies?action=add');
     req.setRequestHeader('Content-Type', 'application/json');
 
     // PARSING RESPONSE SENT FROM SERVER
@@ -203,10 +203,10 @@ document.getElementById('removebutton').addEventListener('click', async function
     var req = new XMLHttpRequest();
     // CREATING JSON OBJECT TO SEND AS POST REQUEST
     var payload = {enemyID: null, regionID: null};
-    payload.enemyID = document.getElementById('removeenemyID').value[0];
-    payload.regionID = document.getElementById('removeregionID').value[0];
+    payload.enemyID = document.getElementById('removeenemyID').value;
+    payload.regionID = document.getElementById('removeregionID').value;
     console.log(payload.enemyID, payload.regionID)
-    req.open('POST', '/RegionEnemies?action=delete');
+    req.open('POST', '/regions/regionEnemies?action=delete');
     req.setRequestHeader('Content-Type', 'application/json');
 
     // PARSING RESPONSE SENT FROM SERVER
